requirements.txt
celery==5.2.3
Django==4.0.2
django-celery-results==2.2.0
djangorestframework==3.13.1